+++ src/main/java/com/example/demo/repository/ActivityLogRepository.java
@@ -3,6 +3,6 @@ import com.example.demo.entity.ActivityLog;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 @Repository
-public interface ActivityLogRepository extends JpaRepository<ActivityLog,Integer>{
+public interface ActivityLogRepository extends JpaRepository<ActivityLog,Long>{
 
 }
