+++ src/main/java/com/example/demo/repository/ActivityTypeRepository.java
@@ -3,6 +3,6 @@ import com.example.demo.entity.ActivityType;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 @Repository
-public interface ActivityTypeRepository extends JpaRepository<ActivityType,Integer>{
+public interface ActivityTypeRepository extends JpaRepository<ActivityType,Long>{
 
 }
