+++ src/main/java/com/example/demo/repository/EmissionFactorRepository.java
@@ -3,6 +3,6 @@ import com.example.demo.entity.EmissionFactor;
 import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 @Repository
-public interface EmissionFactorRepository extends JpaRepository<EmissionFactor,Integer>{
+public interface EmissionFactorRepository extends JpaRepository<EmissionFactor,Long>{
 
 }
